# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArtisanPath API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  name: string;
  phone: string;
  city: string;
  createdAt: string;
}

export interface UserLoginBody {
  name: string;
  phone: string;
  city: string;
}

export interface Artisan {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  subcategory?: string | null;
  city: string;
  availableForTeaching: boolean;
  rating: number;
  verified: boolean;
  createdAt: string;
}

export interface ArtisanLoginBody {
  name: string;
  category: string;
  subcategory?: string;
  city: string;
  availableForTeaching: boolean;
}

export interface Product {
  id: number;
  artisanId: number;
  artisanName: string;
  artisanCity: string;
  title: string;
  description: string;
  /** @nullable */
  descriptionHindi?: string | null;
  /** @nullable */
  descriptionTelugu?: string | null;
  /** @nullable */
  descriptionTamil?: string | null;
  category: string;
  subcategory: string;
  imageUrl: string;
  price: number;
  estimatedPriceMin: number;
  estimatedPriceMax: number;
  city: string;
  averageRating: number;
  reviewCount: number;
  verified: boolean;
  createdAt: string;
}

export interface CreateProductBody {
  artisanId: number;
  title: string;
  description: string;
  category: string;
  subcategory: string;
  imageUrl: string;
  price: number;
  city: string;
}

export type ProductStatsByCategory = { [key: string]: number };

export interface ProductStats {
  topRated: Product[];
  recentlyAdded: Product[];
  byCategory: ProductStatsByCategory;
}

export interface Course {
  id: number;
  artisanId: number;
  artisanName: string;
  artisanCity: string;
  artisanRating: number;
  title: string;
  description: string;
  category: string;
  subcategory: string;
  durationHours: number;
  price: number;
  imageUrl: string;
  createdAt: string;
}

export interface CreateCourseBody {
  artisanId: number;
  title: string;
  description: string;
  category: string;
  subcategory: string;
  durationHours: number;
  price: number;
  imageUrl: string;
}

export interface Event {
  id: number;
  name: string;
  location: string;
  city: string;
  date: string;
  category: string;
  description: string;
  imageUrl: string;
  bookingCount: number;
  maxCapacity: number;
  createdAt: string;
}

export interface BookEventBody {
  userId: number;
  name: string;
  phone: string;
}

export interface RegisterEventBody {
  artisanId: number;
}

export interface EventBooking {
  id: number;
  eventId: number;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  artisanId?: number | null;
  type: string;
  createdAt: string;
}

export interface ChatMessage {
  id: number;
  artisanId: number;
  userId: number;
  senderType: string;
  message: string;
  createdAt: string;
}

export interface SendMessageBody {
  artisanId: number;
  userId: number;
  senderType: string;
  message: string;
}

export interface Rating {
  id: number;
  productId: number;
  userId: number;
  userName: string;
  stars: number;
  review: string;
  createdAt: string;
}

export interface RateProductBody {
  productId: number;
  userId: number;
  userName: string;
  stars: number;
  review: string;
}

export interface Order {
  id: number;
  productId: number;
  userId: number;
  productTitle: string;
  productImageUrl: string;
  price: number;
  status: string;
  createdAt: string;
}

export interface PlaceOrderBody {
  productId: number;
  userId: number;
  price: number;
}

export type ListArtisansParams = {
  city?: string;
  category?: string;
  teachingOnly?: boolean;
};

export type ListProductsParams = {
  category?: string;
  subcategory?: string;
  city?: string;
  artisanId?: number;
};

export type ListCoursesParams = {
  category?: string;
};

export type ListEventsParams = {
  city?: string;
  category?: string;
};

export type ListMessagesParams = {
  artisanId: number;
  userId: number;
};

export type ListOrdersParams = {
  userId?: number;
  artisanId?: number;
};