# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ArtisanPath API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /artisans/login:
    post:
      operationId: loginArtisan
      tags: [artisans]
      summary: Login or register an artisan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtisanLoginBody"
      responses:
        "200":
          description: Artisan logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artisan"
  /artisans:
    get:
      operationId: listArtisans
      tags: [artisans]
      summary: List all artisans
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: teachingOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of artisans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artisan"
  /artisans/{id}:
    get:
      operationId: getArtisan
      tags: [artisans]
      summary: Get artisan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Artisan found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artisan"
        "404":
          description: Not found
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List available courses from artisans who teach
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Courses list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Artisan creates a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Events list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
  /events/{id}/book:
    post:
      operationId: bookEvent
      tags: [events]
      summary: Book an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookEventBody"
      responses:
        "200":
          description: Booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventBooking"
  /events/{id}/register:
    post:
      operationId: registerForEvent
      tags: [events]
      summary: Artisan registers for event to showcase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterEventBody"
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventBooking"
  /chat/messages:
    get:
      operationId: listMessages
      tags: [chat]
      summary: List chat messages for a conversation
      parameters:
        - name: artisanId
          in: query
          required: true
          schema:
            type: integer
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a chat message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
...
  /orders:
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place an order for a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderBody"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for a user
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: artisanId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
components:
  schemas:
...
    Artisan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        subcategory:
          type: ["string", "null"]
        city:
          type: string
        availableForTeaching:
          type: boolean
        rating:
          type: number
        verified:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - category
        - city
        - availableForTeaching
        - rating
        - verified
        - createdAt
    ArtisanLoginBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        subcategory:
          type: string
        city:
          type: string
        availableForTeaching:
          type: boolean
      required:
        - name
        - category
        - city
        - availableForTeaching
...
    Course:
      type: object
      properties:
        id:
          type: integer
        artisanId:
          type: integer
        artisanName:
          type: string
        artisanCity:
          type: string
        artisanRating:
          type: number
        title:
          type: string
        description:
          type: string
        category:
          type: string
        subcategory:
          type: string
        durationHours:
          type: integer
        price:
          type: number
        imageUrl:
          type: string
        createdAt:
          type: string
      required:
        - id
        - artisanId
        - artisanName
        - artisanCity
        - artisanRating
        - title
        - description
        - category
        - subcategory
        - durationHours
        - price
        - imageUrl
        - createdAt
    CreateCourseBody:
      type: object
      properties:
        artisanId:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        subcategory:
          type: string
        durationHours:
          type: integer
        price:
          type: number
        imageUrl:
          type: string
      required:
        - artisanId
        - title
        - description
        - category
        - subcategory
        - durationHours
        - price
        - imageUrl
    Event:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        city:
          type: string
        date:
          type: string
        category:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        bookingCount:
          type: integer
        maxCapacity:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - location
        - city
        - date
        - category
        - description
        - imageUrl
        - bookingCount
        - maxCapacity
        - createdAt
    BookEventBody:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        phone:
          type: string
      required:
        - userId
        - name
        - phone
    RegisterEventBody:
      type: object
      properties:
        artisanId:
          type: integer
      required:
        - artisanId
    EventBooking:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        userId:
          type: ["integer", "null"]
        artisanId:
          type: ["integer", "null"]
        type:
          type: string
        createdAt:
          type: string
      required:
        - id
        - eventId
        - type
        - createdAt
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        artisanId:
          type: integer
        userId:
          type: integer
        senderType:
          type: string
        message:
          type: string
        createdAt:
          type: string
      required:
        - id
        - artisanId
        - userId
        - senderType
        - message
        - createdAt
    SendMessageBody:
      type: object
      properties:
        artisanId:
          type: integer
        userId:
          type: integer
        senderType:
          type: string
        message:
          type: string
      required:
        - artisanId
        - userId
        - senderType
        - message
...
    Order:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        userId:
          type: integer
        productTitle:
          type: string
        productImageUrl:
          type: string
        price:
          type: number
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - productId
        - userId
        - productTitle
        - productImageUrl
        - price
        - status
        - createdAt
    PlaceOrderBody:
      type: object
      properties:
        productId:
          type: integer
        userId:
          type: integer
        price:
          type: number
      required:
        - productId
        - userId
        - price